Rename osint.py to footprints.py
