+++ scanners/footprints.py
